basic/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "basic"
version = "0.1.0"
description = "Minimal starter workflow project"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "llama-index-core>=0.14.5",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-llms-openai-like>=0.5.3",
    "llama-index-vector-stores-weaviate>=1.4.1",
    "llama-index-workflows>=2.5.0,<3.0.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "weaviate-client>=4.17.0",
]

[dependency-groups]
dev = [
    "hatch>=1.14.2",
    "pytest>=8.4.2",
    "ruff>=0.13.2",
    "ty>=0.0.1a21",
]

[tool.hatch.envs.default.scripts]
format = "ruff format ."
format-check = "ruff format --check ."
lint = "ruff check --fix ."
lint-check = ["ruff check ."]
typecheck = "ty check src"
test = "pytest"
all-check = ["format-check", "lint-check", "test"]
all-fix = ["format", "lint", "test"]

[tool.llamadeploy]
env_files = [".env"]

[tool.llamadeploy.workflows]
default = "basic.workflow:wf"

[tool.llamadeploy.ui]
directory = "ui"
